+++ lib/ga/index.tsx
@@ -1,0 +1,1 @@
+export { ga } from './ga';
